added google auth and login page

+++ src/app/api/db/route.ts
@@ -1,19 +1,28 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
-import { NextResponse } from "next/server";
+import { NextRequest, NextResponse } from "next/server";
 import { createClient } from "../../lib/supabaseClient";
 import { ClientStockData } from "@/app/lib/types";
 
 // return list of stocks saved in supabase db
-export async function GET(): Promise<
-  NextResponse<ClientStockData[] | { error: string }>
-> {
+export async function GET(
+  req: NextRequest
+): Promise<NextResponse<ClientStockData[] | { error: string }>> {
   const supabase = await createClient();
+
+  // parse the URL
+  const { searchParams } = new URL(req.url);
+  const userId = searchParams.get("userId"); // reads ?userId=<uid>
+
+  if (!userId) {
+    return NextResponse.json({ error: "Missing userId" }, { status: 400 });
+  }
 
   // load all stocks
   const { data, error } = await supabase
     .from("stocks")
     .select("*")
+    .eq("user_id", userId)
     .order("quantity", { ascending: false });
 
   let stockData: ClientStockData[] = [];
@@ -43,12 +52,13 @@ export async function POST(req: Request): Promise<NextResponse> {
     const supabase = await createClient();
 
     // Parse the JSON body from the request
-    const { name, wealth_simple, quest_trade, dividend_frequency } =
+    const { name, wealth_simple, quest_trade, dividend_frequency, user_id } =
       await req.json();
 
     // Insert into Supabase, mapping request fields to DB columns
     const { data, error } = await supabase.from("stocks").insert([
       {
+        user_id: user_id,
         stock_name: name,
         wealth_simple: wealth_simple,
         quest_trade: quest_trade,
@@ -72,7 +82,7 @@ export async function PUT(req: Request): Promise<NextResponse> {
     const supabase = await createClient();
 
     // Parse JSON body
-    const { name, wealth_simple, quest_trade, dividend_frequency } =
+    const { name, wealth_simple, quest_trade, dividend_frequency, user_id } =
       await req.json();
 
     if (!name) {
@@ -86,6 +96,7 @@ export async function PUT(req: Request): Promise<NextResponse> {
     const { data, error } = await supabase
       .from("stocks")
       .update({
+        user_id,
         stock_name: name,
         wealth_simple,
         quest_trade,
@@ -106,7 +117,7 @@ export async function DELETE(req: Request): Promise<NextResponse> {
   try {
     const supabase = await createClient();
     // Parse the JSON body from the request
-    const { stock_name } = await req.json();
+    const { stock_name, user_id } = await req.json();
 
     if (!stock_name) {
       return NextResponse.json(
@@ -118,7 +129,8 @@ export async function DELETE(req: Request): Promise<NextResponse> {
     const { data, error } = await supabase
       .from("stocks")
       .delete()
-      .eq("stock_name", stock_name);
+      .eq("stock_name", stock_name)
+      .eq("user_id", user_id);
 
     if (error) {
       throw error;
